// src/pages/wms/station/instances/replenish/type.ts
export enum StationOperationType {
    conveyorArea = "conveyorArea",
    robotArea = "robotArea",
    selectDetailArea = "selectDetailArea",
    shelfArea = "shelfArea",
    collectingDoc = "collectingDoc",
    callBin = "callBin",
    infoInput = "infoInput",
    scanInfoListArea = "scanInfoListArea",
    defaultArea = "defaultArea",
    tips = "TIPS",
    orderArea = "orderArea"
}